amavis: Pass on uncheckable encrypted content by default
Previously we would report encrypted archives either as viruses (ieeek)
or, more recently, reject them because analysis failed. For such
content, however, it is normal that analysis fails because we do not
know the credentials to open the encryption. AMaViS already has logic to
detect that content is encrypted but no easy way to totally exclude such
content from analysis. What we can do however is to distinguish based on
content category after analysis has failed whether the content is
encrypted or not and therefore pass it on or not.