Add cuckoo process instance systemd unit
Now that we've disabled results processing by the main cuckoo process we
need to start separate processors. While our cuckooprocessor.sh is doing
just that for embed mode we need an actual systemd unit for api mode.
Based on cuckoosandbox/cuckoo#2025 we use an instantiatable template
unit to start a configurable number of processor instances.

Closes #20 (again :).