Deliver Peekaboo without git submodule
Switch from a git submodule back to pip for delivery of the actual
Peekaboo distribution because the submodule approach has a number of
drawbacks. Implement the initial goal of allowing tests of unreleased
Peekaboo versions by looking for a manually provided sourc directory as
suggested and discussed in issue #9.

Requires a Peekaboo distribution which installs ruleset.conf.sample into
<venv>/share/doc/peekaboo.

Closes #35.
Closes #36.