Start Peekaboo after finished amavis installation
There was an error when peekaboo is started before amavis files, user, group ...
is installed.

When installing peekaboo the service is restarted when finished.
At this point the user amavis doesn't exist because the next play
then installs it. By moving the restart to after the amavis
play the error doesn't occur and the playbook continues.