cuckoo: Run services as separate user
With API mode there is no longer any reason to install and run cuckoo
services as the peekaboo user. Since we're installing in API mode by
default and upstream (us, Peekaboo) is in the process of deprecating and
removing embedded mode, separate cuckoo services out with their own
cuckoo user. This should also remove the last interdependency between
cuckoo and peekaboo requiring them to run on the same machine, as can be
seen by the creation of the peekaboo user moving down to after cuckoo
installation in the playbook.