systemd: Make mysql a required service for cuckoo
The cuckoo main services can't run without mysql. So make it a hard
requirement in the systemd units so that systemd starts it automatically
if one of those services is started.

We're ignoring postgresql here because this installer does not support
postgres. An admin can still make it work with postgres using suitable
local systemd drop-in configs.