Remove unnecessary prerequisite packages
python-setuptools and python-pip are not required because the virtualenv
command from python{,3}-virtualenv automatically installs them into the
new venv.

Install mysqlclient python package explicitly because the mysql meta
package would pull in MySQL-python for python2 and mysqlclient for
python3. The former is unmaintained since 2014 but does not need
libmysqlclient-dev installed. We prefer the latter for consistency which
supports python 2 and 3 but needs libmysqlclient-dev installed. But we
did install that anyway for the python3 install option of peekaboo.

sqlite3 was never strictly required by anything.