tcpdump: Restrict access and install on right machine
Move the installation of tcpdump down to after the cuckoo user has been
created so it can be owned by the cuckoo group. Tweak file permissions
so that only the cuckoo group is allowed to execute it. This, together
with the existing file capabilities configuration allows the cuckoo user
and only the cuckoo user to use it without root privileges.

Fix the host assignment to be the cuckoo server while at it.

Closes #52.