cuckoo: Work around flask dependency problem
Work around flask not pinning its dependency versions, pulling in a
werkzeug that is too new for itself. We rely on pip's laziness in
upgrading to leave this version alone as long as it suffices for flask.
Should become irrelevant as cuckoo switches to a newer flask which
supports and pulls in a newer werkzeug. We do this in a separate step to
avoid conflicts within a single transaction.

Closes #57.