ubuntu: Prevent apt-get from asking questions
apt-get dist-upgrade run from the script croaks when upgrading grub
because it spawns a dialog asking for the device to reinstall grub to.

Switch apt-get to the from asking questions because we want to be an
automated installer, particularly because we do want to work when run
from a Dockerfile or Vagrantfile.

Closes #54.