amavis: Reorder group membership change and service start
We had a delayed start of the peekaboo service so the user had been made
part of the amavis group at that point so it could read file created by
amavis. But actually the group membership change happened even later
than that. This was masked by our starting of the service with *primary*
group amavis by the systemd unit.

Move the group membership change up so the service can work right away
even without systemd.