peekaboo: Explicitly secure socket
After switching to using the supplementary group list of the peekaboo
user for client file access, we need to explicitly open up our socket to
those clients for submit.

Sync up the peekaboo config file with recent peekaboo that has gained
support for explicit socket access config. We would now normally create
a separate peekaboo-client group as recommended by upstream. But amavis
does not support initgroup()ing to its users supplementary group list
and would not be able to access the socket even if it were member of the
clients group. As a workaround we use the amavis primary group as socket
owner since currently we only have one actual client, amavis.