Adjust for Ubuntu 20.04
Ubuntu 20.04 all but drops support for python2.7. All respective
packages apart from the interpreter itself are removed. This
necessitates some adjustments to our installer.

Peekaboo in current development has already dropped python2 support. So
we can drop it in the installer as well. We remove installation of
python2 packages which no longer exist, switch to python3 versions of
packages we still need and remove the switch of the installer to select
which python interpreter to use for Peekaboo.

To get the virtualenv command we need to install differing packages
because while python2 and 3 versions of the package exist in old and
current Ubuntu releases, only one provides the actual virtualenv command
in /usr/bin and this has switched from python-virtualenv to
python3-virtualenv with 20.04.

Ansible in Ubuntu 20.04 still defaults to using python2 as the
interpreter run on remote machines for acutal task execution which no
longer works with 20.04 because prerequisite modules are missing. We
force this to be python3 instead which works on 18.04 as well.

Cuckoo however still has a hard dependency on python2.7. It installs
okay into its virtualenv but is otherwise untested.