vagrant: Exercise the whole installer
Up until now we used the installer script to install the ansible
prerequisites and then ran the vagrant ansible_local provisioner
explicitly. This allowed for better error handling, more control and
(not the least :) coloured output. But it did not exercise the actual
installer. Also it produced some confusing output with a warning that
ansible was not run only to then run ansible.

Upcoming changes will change the way ansible is run (with python3 forced
as interpreter). This showcases that with the current scheme our
Vagrantfile is not actually exercising the installer as a whole but
something similar that can get out of sync and simply is
non-representative.

So this change removes that distinction and runs the installer as is
from the Vagrantfile. As a side-effect we can now remove the NOANSIBLE
logic and warning from the installer script.