Allow MariaDB server to be separate machine
We've had a mariadb_server inventory entry for a long time in order to
dynamically configure Peekaboo and Cuckoo to use it. But the MariaDB
server we installed always ran on the peekabooav_server. There's no
reason for that, so switch it. And remove a currently superfluous and
lonely tag from the play until a reason to have those becomes apparent.