peekaboo: Updated Dockerfile
This remakes the Dockerfile for peekaboo, with a proper multi stage build.
The first container builds peekaboo and the built application is then
copied to the final container.

It also purges what packages we install in each container. This together
with other 'tricks' is used to keep the image size small.

Also the placeholder replacement and application start is moved to an
entrypoint script for easier maintainability.