rspamd: Added rspamd container and needed files
This patches the peekaboo module into rspamd and configures rspamd for a
functional showcase, with the help of an entrypoint script.

The entrypoint inspects an environment variable RSPAMD_ENABLED_MODULES.
If it exists, it walks through all .lua files in
/usr/share/rspamd/plugins and disables all modules but the ones listed
in the environment variable via enabled=false statements in
/etc/rspamd/rspamd.conf.local. The enabled modules are switched on
explicitly using enabled=true statements (required e.g. for module dcc).

These mechanics imply that all modules have a .lua file in the plugins
directory which seems to be the case for all but fuzzy_check. The latter
is disabled by default because it's unconfigured.

Module trie is handled specially because it expects a symbol table and
has no parameter enabled.