cortex: Make API URL configurable and generate token
Our changes for Docker and docker-compose introduced placeholders
cortex_url and cortex_api_token to Peekaboo's analyzers.conf. The
Ansible installer chokes on these (undefined) placeholders.

So we define variables for both to make the installer work again. As a
side-effect, this makes the Cortex API URL configurable and provides a
generated, secret API token in the config. The API token needs to be
configured in Cortex as well, which makes our generating it somewhat
useless, currently.

In the default configuration, Cortex is not used. The mere presence of
these values in analyzers.conf does not change behaviour and thus does
not hurt.