Add GitHub Actions CI
Add a GitHub Actions workflow that runs the installer on Ubuntu 18.04
and 20.04 for validation and regression testing. We run the installer
directly on the GitHub Ubuntu runner systems. Because they're heavily
customised, we first need to undo some of that customisation so our
Installer can succeed. So we're effectively testing our installer under
quite adverse conditions.