pipeline: Modify swaks attachment commands for older versions
Switch the swaks commands to an --attach syntax that works with both
legacy and current versions by avoiding reading from a file and instead
supplying the input on stdin. Add a note explaining the reason and
further notes why the @ can be used with the postfix_tx
container-supplied version. While at it, add an explanation how
docker-compose exec coupled with stdin can be used to pipe attachment
content to the containerised swaks from the outside.