cortex setup: Use pwgen to generate password
We can use pwgen to auto-generate a password for the Cortex admin to
make it clearer what's happening and avoid shooting ourselves in the
foot by doing it wrong (even more than auto-generating a password ;). So
we add it to the container image and use it in the script file.