cortex setup: Improve argument handling
Numeric comparison does not need the extended features of [[, so we use
a standard test with [, gaining improvided POSIX compatibility.

$1 can be empty an needs to be quoted. For paranoia's sake we go ahead
and quote $# as well.

Some shells (particularly dash, again) throw an error when trying to
shift past the end of the argument list. So we rework the argument
shifting and check for empty arguments. This also adds helpful
diagnostic messages for the user.