cortex setup: Avoid escaping-interpreting echo
Behaviour of the echo command regarding interpretation of escape
sequences and support of the -e option is inconsistent between shells,
dash again highlighting the fact. printf seems to be universally
available and well-defined. In order to aid debugging of the script on
dash-using systems we switch to printf in favour of echo -e.