cortex setup: Handle additional error conditions
If cortex is not reachable, curl will output an http_code of 000. The
exit code is more helpful with a plethora of different values for
different error cases. For our use-cases (particularly as setup job in
k8s) it's enough to detect connection errors through the return code and
exit.

Because the output of curl can potentially be empty as well (e.g. if the
command were to be not found), we quote its use throughout the script.