peekaboo: Support different flavours of container image
Modify the Dockerfile so that usage of a locally-supplied PeekabooAV
source directory becomes optional. This is implemented by way of
globbing for the source file on COPY and later checking directory
existance. If it doesn't exist, we fall back to installing the latest
release from PyPI, optionally requesting a specific version governed by
the VERSION build argument.

This brings the Dockerfile in line with the behaviour of our Ansible
installer and allows for building versioned release images based on the
actual release on PyPI, not a local git checkout.