k8s: Make Cortex analyses work on k8s
Switch to pre-build images from Docker Hub, particularly one for Cortex
that does contain the kubernetes job runner from a pending PR (link not
given here as to not spam it when pushing this commit multiple times).

Add the required shared persistent volume claim to the specs and
configs. Together with a service account that is allowed to manage jobs
this allows to run analyses on k8s.