ansible: Pin PyPI version to current release
Limit the PeekabooAV version installed from PyPI to be lower than 2.2.
This would install the latest point release of our current release 2.1.

The goal here is to have point releases of Peekaboo and the Installer be
independent of each other so we can do fixes that require a version bump
in both without confusing users too much. Always installing the latest
point relase (by default) seems a good tradeoff.

By making the pip constraint a variable in group_vars/all.yaml makes
it fully overrideable on the Ansible-level, e.g. to ==2.1.1 if desired.
It can also be circumvented by locally supplying a PeekabooAV directory
to the installer.

An explanatory note is added to the README as well.

Note: Pre-releases are excluded anyway and not installed unless
explicitly requested as peekabooav==2.1rc3 or by enabling the --pre flag
to pip install.