rspamd, postfix: Switch to alpine latest
For alpine, the latest tag points to the latest stable release
(currently 3.15). We can use that to simplify our Dockerfiles by basing
our images on the latest tag. That way we always get the latest stable
release of alpine when building our images. That way we get alerted to
regressions early on when new releases happen. The argument of stability
and reproducibility fielded by proponents of version-pinning is already
underminded by the fact that we're not pinning alpine package versions,
revisions and hashes either. So we can just as well go with the flow and
sort out breakage as it happens without the need of tracking and
switching stable distro releases as well.