ci: Add scheduled build and push of container images
Regularly rebuilding all (supported) container images helps catch
regressions through external dependencies and keeps them up-to-date with
security patches. Therefore we add a scheduled Github Actions workflow
that does just that. It's using a dynamic matrix build that determines
the refs to build based on a filter list applied to the Installer repo
refs.

Unfortunately, the docker metadata action does not cater well to image
rebuilds. It basically assumes all releases to monotonically increase
versions and never being processed simulataneously or out of order.  It
also assumes to be run on the source code corresponding to the
github.ref the workflow was called upon. This is only true for PRs,
pushes and taggings but not for scheduled builds.

Therefore we use a patched version of the action that accepts the ref to
work with as an input. Additionally we determine the latest version and
latest release per major release in preparation of the scheduled matrix
build and provide those the action where it influences the tagging
scheme.