pipeline: Simplify postfix config
Similarly to how we're doing it in Kuberntes we can simplify the postfix
configuration by exploiting the fact that the systems know each other
by their names by way of the compose network name resolution. Therefore
we can tell the client talking to postfix-tx to send a mail to
root@postfix-rx directly. We only need to allow this unauthenticated
relaying on postfix-tx. This allows us to get rid of the routing from
postfix-tx's root account to postfix-rx through the virtual table.